requirements.txt
mediapipe>=0.10.0
opencv-python>=4.6.0
pymongo[srv]
pyaudio>=0.2.11
requests>=2.28.1
elevenlabs>=0.0.1
pydub>=0.25.1
PyQt5>=5.15.0
pyautogui>=0.9.53
pymongo[srv]
google-generativeai
